+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Updates
 - Moved `Base16::toRawString` to `MineCommon::byteArrayToRawString`
 - Added `MineCommon::rawStringToByteArray`
+- Renamed `BigIntegerHelper` to `MathHelper`
 
 ## [1.0.0] - 05-09-2017
 ### Added